# docs/10_api_openapi.yaml
openapi: 3.0.3
info:
  title: Freight Forwarding and Finance Platform API
  version: 2.0.0
  description: >-
    Organization- and branch-scoped API for freight operations, informational
    service charges, reusable financial documents, and double-entry accounting.
servers:
  - url: /api/v1
    description: Current API
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Authentication]
      security: []
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200': {description: Authenticated, content: {application/json: {schema: {$ref: '#/components/schemas/AuthResponse'}}}}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '429': {description: Too many attempts}
...
  /auth/refresh:
    post:
      tags: [Authentication]
      security: []
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefreshRequest'}
      responses:
        '200': {description: Token refreshed, content: {application/json: {schema: {$ref: '#/components/schemas/AuthResponse'}}}}
  /organizations:
    get:
      tags: [Organizations]
      operationId: listOrganizations
      responses: {'200': {description: Organizations, content: {application/json: {schema: {$ref: '#/components/schemas/PagedOrganizationResponse'}}}}}
    post:
      tags: [Organizations]
      operationId: createOrganization
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/OrganizationCreate'}}}
      responses: {'201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Organization'}}}}}
  /organizations/{organizationId}/branches:
    get:
      tags: [Organizations]
      operationId: listBranches
      parameters: [{$ref: '#/components/parameters/OrganizationId'}]
      responses: {'200': {description: Branches, content: {application/json: {schema: {$ref: '#/components/schemas/PagedBranchResponse'}}}}}
    post:
      tags: [Organizations]
      operationId: createBranch
      parameters: [{$ref: '#/components/parameters/OrganizationId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/BranchCreate'}}}
      responses: {'201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Branch'}}}}}
  /users:
    get:
      tags: [Users and Authorization]
      operationId: listUsers
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {description: Users, content: {application/json: {schema: {$ref: '#/components/schemas/PagedUserResponse'}}}}}
    post:
      tags: [Users and Authorization]
      operationId: inviteUser
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/UserCreate'}}}
      responses: {'201': {description: User invited, content: {application/json: {schema: {$ref: '#/components/schemas/User'}}}}}
  /users/{userId}/role-assignments:
    post:
      tags: [Users and Authorization]
      operationId: assignRole
      parameters: [{$ref: '#/components/parameters/UserId'}]
      requestBody:
        required: true
        content: {application/json: {schema: {$ref: '#/components/schemas/RoleAssignmentCreate'}}}
      responses: {'201': {description: Role assigned}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    OrganizationId: {name: organizationId, in: path, required: true, schema: {type: integer, format: int64}}
    UserId: {name: userId, in: path, required: true, schema: {type: integer, format: int64}}
...
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
    PageSize: {name: page_size, in: query, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
...
  responses:
    Unauthorized: {description: Authentication required}
...
  schemas:
...
    PageMeta:
      type: object
      properties: {page: {type: integer}, page_size: {type: integer}, total: {type: integer}}
    LoginRequest:
      type: object
      required: [username, password]
      properties: {username: {type: string}, password: {type: string, format: password}}
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties: {refresh_token: {type: string}}
    AuthResponse:
      type: object
      properties: {access_token: {type: string}, refresh_token: {type: string}, token_type: {type: string, example: Bearer}, expires_in: {type: integer}}
    OrganizationCreate:
      type: object
      required: [organization_code, legal_name, default_currency_code]
      properties: {organization_code: {type: string}, legal_name: {type: string}, display_name: {type: string}, default_currency_code: {type: string, minLength: 3, maxLength: 3}, timezone: {type: string}}
    Organization:
      allOf: [{ $ref: '#/components/schemas/OrganizationCreate' }, {type: object, properties: {id: {type: integer}, status: {type: string}}}]
    BranchCreate:
      type: object
      required: [branch_code, name]
      properties: {branch_code: {type: string}, name: {type: string}, place_id: {type: integer}, address: {type: string}, is_head_office: {type: boolean}}
    Branch:
      allOf: [{ $ref: '#/components/schemas/BranchCreate' }, {type: object, properties: {id: {type: integer}, organization_id: {type: integer}}}]
    UserCreate:
      type: object
      required: [user_code, username, email, display_name]
      properties: {user_code: {type: string}, username: {type: string}, email: {type: string, format: email}, display_name: {type: string}}
    User:
      allOf: [{ $ref: '#/components/schemas/UserCreate' }, {type: object, properties: {id: {type: integer}, status: {type: string}}}]
    RoleAssignmentCreate:
      type: object
      required: [role_id, organization_id]
      properties: {role_id: {type: integer}, organization_id: {type: integer}, branch_id: {type: integer, nullable: true}, starts_at: {type: string, format: date-time}, expires_at: {type: string, format: date-time, nullable: true}}
...
    PagedOrganizationResponse: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Organization'}}, meta: {$ref: '#/components/schemas/PageMeta'}}}
    PagedBranchResponse: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/Branch'}}, meta: {$ref: '#/components/schemas/PageMeta'}}}
    PagedUserResponse: {type: object, properties: {items: {type: array, items: {$ref: '#/components/schemas/User'}}, meta: {$ref: '#/components/schemas/PageMeta'}}}
...